Move Estimator to estimators package
